refactor hook structure
